refactor(layout): thin package exports to compute_layout
Drop __all__ and convenience re-exports; import types and extract
helpers from the defining submodules.
